requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Unidecode == 1.2.0
dask == 2.30.0
h5py == 3.1.0
implicit == 0.4.4
ipywidgets == 7.5.1
matplotlib == 3.3.2
numpy == 1.19.5
pandas == 1.1.3
plotly == 4.14.3
scikit_learn == 0.23.2
scikit_surprise == 1.1.1
scipy == 1.4.1
seaborn == 0.11.0
spotipy == 2.18.0
tqdm == 4.43.0
yellowbrick == 1.3.post1
